fix: viewModel 연결 오류 해결
